Add oneapi compiler
This PR adds the oneapi and dpcpp compilers to archspec. These compilers
are Intel's latest generation compilers and have similar options to the
classic Intel compilers. As such, the entries here are based on the
intel entires.

Note: I also added entries, including for intel, to the zen3 block. If
there was a reason that the intel compiler was not there then the new
entries in that section will need to be removed.